+++ admin/js/woo-amz-integration-admin.js
@@ -21,7 +21,7 @@
 		
 						if ( feedEnabled ) {
 			
-							tfsWooAmzAjax( feedEnabled );
+							tfsWooAmzAjax( feedEnabled, false, true );
 	
 							$('#send_inventory').hide();
 	
@@ -35,16 +35,48 @@
 							$('#download_inventory').hide();
 	
 							$('#feed-warning').show();
+
+							$('#feed_submit').hide();
+							$('#feed_continue').hide();
 	
-			
 						} else {
 		
-							alert('Enable Feed and Save before running!');
+							alert('Enable Feed and click "Save Settings" before running!');
 		
 						}
 			
 					});
+
+					if ( $( '#feed_continue' ) ) {
+
+						$( '#feed_continue' ).click(function() {
+
+							if ( feedEnabled ) {
 	
+								tfsWooAmzAjax( feedEnabled, true, false );
+	
+								$('#send_inventory').hide();
+		
+								$('#feed-progress').show();
+		
+								$('#feed-progress').html('<p>Current Progress: <strong>Loading...</strong></p>');
+		
+								$('#feed-status').show();
+								$('#feed-status-text').html('<p>Current Status: <strong>Starting</strong></p>');
+		
+								$('#download_inventory').hide();
+		
+								$('#feed-warning').show();
+								
+								$('#feed_continue').hide();
+								$('#feed_submit').hide();
+	
+							}
+	
+						});
+	
+					}
+
 				}
 
 			}
@@ -57,7 +89,7 @@
 	tfsInitAdminPage();
 
 
-	function tfsWooAmzAjax( value, timer = null ) {
+	function tfsWooAmzAjax( value, runFeed, restartFeed, timer = null ) {
 
 		console.log('okay');
 
@@ -71,7 +103,9 @@
 				method: 'GET',
 				url: tfs_woo_amz_int_object.api_url + 'woo-amz-feed/',
 				data: {
-				  enabled: value
+				  enabled: value,
+				  run: runFeed,
+				  restart: restartFeed
 				},
 				beforeSend: function ( xhr ) {
 
@@ -88,21 +122,37 @@
 				complete: function( data ) {
 
 					let parsed_data = JSON.parse( data.responseJSON );
-	
-					console.log('complete: ' + parsed_data );
 
+					console.log( parsed_data );
+
+	
 					//try to see if parsed_data.completed is available yet, if not, the feed is new so start it
 					try {
 
-						if ( parsed_data.completed == 0 ) {
+						if ( parsed_data['status'].completed == 0 ) {
 		
-							$('#feed-progress').html('<p>Current Progress: '+ parsed_data.products_processed + ' / ' + parsed_data.products_to_process + '</p>');
+							$('#feed-progress').html('<p>Current Progress: '+ parsed_data['status'].products_processed + ' / ' + parsed_data['status'].products_to_process + '</p>');
 		
 							$('#feed-status-text').html('<p>Current Status: <strong>Running...</strong></p>');
+
+							console.log( parsed_data );
+
+							/**
+							 * If the feed was restarted, change the data property values so that
+							 * the feed will continue on the next REST API call and not restart again
+							 */
+							if ( parsed_data.continue_after_reset == true ) {
+
+								runFeed = true;
+								restartFeed = false;
+
+								console.log('hit');
+
+							}
 							
 							timer = setTimeout(worker, 5000);
 		
-						} else if ( parsed_data.completed == 1 ) {
+						} else if ( parsed_data['status'].completed == 1 ) {
 				
 							$('#feed-progress').hide();
 		
@@ -132,7 +182,7 @@
 	
 				},
 				fail: function( data ) {
-	
+
 					console.log('fail: ' + JSON.parse(data.responseJSON) );
 	
 				},